refactor : restructuring the project
